requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.5
flake8==6.0.0
idna==3.4
mccabe==0.7.0
ndg-httpsclient==0.5.1
pep8==1.7.1
Pillow==9.4.0
psycopg2==2.9.5
pusher==3.3.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyNaCl==1.5.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
uWSGI==2.0.21
whitenoise==6.3.0
google-api-python-client==2.97.0
